renamed index to util, going to work on getting example to work in front of working on edge cases. Going totake some general notes on it
