+++ example/cds-hooks/better-patient-view.ts
@@ -1,5 +1,5 @@
-import { Service } from "../../src/cds-hooks";
-import { processPatientNames, buildPatient } from "./index";
+import { Service, Card } from "../../src/cds-hooks";
+import { processPatientNames } from "./util";
 
 /**
  * patient-view-kitchen-sink
@@ -44,16 +44,61 @@ export default new Service(
     }>
   ) => {
     const data = request.prefetch;
-    // const patientNames = processPatientNames(data.patient);
+    const patientNames = processPatientNames(data.patient);
     // @todo no explicit any here, as fhir4.BundleEntry does not outline start and end period for an encounter
     const encounters: Array<any> = [];
     data.encounter.entry?.forEach((entry) => {
       encounters.push(entry);
     });
     // @todo to fill in all of the details for patient, call this
-    const cards = buildPatient(data.patient);
+    // const cards = buildPatient(data.patient);
     return {
-      cards: cards,
+      cards: [
+        // Name(s)
+        new Card({
+          detail: `This patient has ${patientNames.length} name${
+            patientNames.length <= 1 ? "" : "s"
+          } on record.`,
+          source: {
+            label: "Automate Medical, Inc.",
+            url: "https://www.automatemedical.com/",
+          },
+          summary: `Now seeing: ${patientNames[0].given} ${patientNames[0].family}.`,
+          indicator: "info",
+        }),
+        // DOB
+        new Card({
+          source: {
+            label: "Automate Medical, Inc.",
+            url: "https://www.automatemedical.com/",
+          },
+          summary: `Date of birth: ${data.patient.birthDate}`,
+          indicator: "info",
+        }),
+        // DOB
+        new Card({
+          source: {
+            label: "Automate Medical, Inc.",
+            url: "https://www.automatemedical.com/",
+          },
+          summary: `Active: ${data.patient.active === true ? "yes" : "no"}`,
+          indicator: "info",
+        }),
+        // Information on the last encounter
+        new Card({
+          detail: `Last visit was on ${
+            encounters.pop().resource.period.start
+          }. There are ${encounters.length} encounter${
+            encounters.length <= 1 ? "" : "s"
+          } on record.`,
+          source: {
+            label: "Automate Medical, Inc.",
+            url: "https://www.automatemedical.com/",
+          },
+          summary: `Last visit`,
+          indicator: "info",
+        }),
+      ],
     };
   }
 );
