+++ example/cds-hooks/util.ts
